+++ platform-sdk/swirlds-config-extensions/src/main/java/module-info.java
@@ -7,6 +7,9 @@ module com.swirlds.config.extensions {
 
     requires transitive com.swirlds.config.api;
     requires com.swirlds.base;
+    requires com.fasterxml.jackson.core;
+    requires com.fasterxml.jackson.databind;
     requires org.apache.logging.log4j;
+    requires org.yaml.snakeyaml;
     requires static transitive com.github.spotbugs.annotations;
 }
