+++ platform-sdk/swirlds-config-extensions/src/test/resources/test.yaml
@@ -10,3 +10,8 @@ gossip:
   randomList:
     - "foo"
     - "bar"
+  randomListOfList:
+    - - "foo"
+      - "bar"
+    - - "baz"
+      - "qux"
